requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.1
psutil==5.9.5
aiohttp==3.9.1
Pillow>=10.0.0
motor==3.3.2

# Notes:
# - Removed the third-party 'asyncio' package (stdlib module; installing it can cause issues on modern Python).
# - PyNaCl is optional and only needed for voice features. To enable voice, add: PyNaCl
